Add the Spring IO plugin
Configure the Spring IO plugin such that it's only applied when the
build is run with -PplatformVersion=<version>. This platformVersion
property is used to determine the version of the Platform that will
be used when running the springIoCheck task. The plugin can be used
by running a build as follows:

./gradlew clean springIoCheck -PplatformVersion=1.0.0.BUILD-SNAPSHOT -PJDK7_HOME=… -PJDK8_HOME=…

This will test the project on JDK7 and JDK 8 using the dependencies
defined in the latest snapshot of Spring IO Platform 1.0.0.

Running the tests against JDK 8 identified a problem with
CommunicationTemplateTest. The tests relied on the ordering of content
in the JSON payload and that ordering is not guaranteed. To overcome
this the use of direct String comparison has been replaced with JSON
path-based assertions